\name{.Introduction}
\alias{Introduction}

\title{Introduction}
\details{ Library mixed.model.gwas contains functions for performing mixed-model genomewide association analyses. The two unique features of this library are that it handles both SNP- and haplotype-based tests of association, and implements a permutation test for genomewide significance that is consistent with the mixed model control of unequal relatedness. It is optimised for use with the output of the STITCH R package for genotype imputation, although it will work with other imputed datasets if formatted correctly.}

